Make caves turn off.
"NOTE: Grasslands mod modifies map generation settings 'caves' and 'dungeons'. This mod tries to reset these settings when the world is closed, but if they are not (computer or world crash) then the disabled settings will modify the generation of new caves and dungeons in other worlds. To turn on dungeons and caves then open a world with this mod enabled and then leave, or go to settings page and search 'mg_flags' (without quotes) to manually reset the flags. USE AT YOUR OWN RISK!"